レジストリまわりを umapita_registry.cpp に分離した
